Add Dockerfile and docker-compose.yaml for Python Bitcask server
This commit adds a Dockerfile and a docker-compose.yaml file to the
project, which can be used to easily run a Python Bitcask server in a
containerized environment. The Dockerfile specifies the base image to
use and sets up the working directory and entrypoint for the container,
while the docker-compose.yaml file defines the service and ports to
expose. The volumes and environment variables are currently commented
out, but can be uncommented and customized as needed.